Add projects for separated client and server only API
- Add new projects and nuspecs to publish separated NuGet packages with client only and server only APIs.
- Split and extract some exiting classes/properties for convinience.
- Update Azure Pipelines yaml accordingly.
