+++ source/nanoFramework.System.Net.Http-server.nuspec
@@ -1,9 +1,9 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <package xmlns="http://schemas.microsoft.com/packaging/2012/06/nuspec.xsd">
   <metadata>
-    <id>nanoFramework.System.Net.Http</id>
+    <id>nanoFramework.System.Net.Http-server</id>
     <version>$version$</version>
-    <title>nanoFramework.System.Net.Http</title>
+    <title>nanoFramework.System.Net.Http-server</title>
     <authors>nanoFramework project contributors</authors>
     <owners>nanoFramework project contributors</owners>
     <requireLicenseAcceptance>false</requireLicenseAcceptance>
@@ -14,11 +14,10 @@
     <iconUrl>https://secure.gravatar.com/avatar/97d0e092247f0716db6d4b47b7d1d1ad</iconUrl>
     <copyright>Copyright (c) 2018 The nanoFramework project contributors</copyright>
     <description>
-      This package includes the nanoFramework.System.Net.Http assembly for nanoFramework C# projects.
-      There are also two other packages with just the client API and just the server API.
-      These are meant to be used when there is the need to use smaller assemblies.
+      This package includes the nanoFramework.System.Net.Http assembly (server API only) for nanoFramework C# projects.
+      There is also a package with the client API only and another with the full API.
     </description>
-    <summary>nanoFramework.System.Net.Http assembly for nanoFramework C# projects</summary>
+    <summary>nanoFramework.System.Net.Http assembly (server API only) for nanoFramework C# projects</summary>
     <tags>nanoFramework C# csharp netmf netnf nanoFramework.System.Net.Http</tags>
     <dependencies>
       <dependency id="nanoFramework.CoreLibrary" version="[1.1.1]" />
@@ -27,10 +26,10 @@
     </dependencies>
   </metadata>
   <files>
-    <file src="nanoFramework.System.Net.Http\bin\Release\System.Net.Http.dll" target="lib\System.Net.Http.dll" />
-    <file src="nanoFramework.System.Net.Http\bin\Release\System.Net.Http.pdb" target="lib\System.Net.Http.pdb" />
-    <file src="nanoFramework.System.Net.Http\bin\Release\System.Net.Http.pdbx" target="lib\System.Net.Http.pdbx" />
-    <file src="nanoFramework.System.Net.Http\bin\Release\System.Net.Http.pe" target="lib\System.Net.Http.pe" />
-    <file src="nanoFramework.System.Net.Http\bin\Release\System.Net.Http.xml" target="lib\System.Net.Http.xml" />
+    <file src="nanoFramework.System.Net.Http-server\bin\Release\System.Net.Http.dll" target="lib\System.Net.Http.dll" />
+    <file src="nanoFramework.System.Net.Http-server\bin\Release\System.Net.Http.pdb" target="lib\System.Net.Http.pdb" />
+    <file src="nanoFramework.System.Net.Http-server\bin\Release\System.Net.Http.pdbx" target="lib\System.Net.Http.pdbx" />
+    <file src="nanoFramework.System.Net.Http-server\bin\Release\System.Net.Http.pe" target="lib\System.Net.Http.pe" />
+    <file src="nanoFramework.System.Net.Http-server\bin\Release\System.Net.Http.xml" target="lib\System.Net.Http.xml" />
   </files>
 </package>
