+++ source/nanoFramework.System.Net.Http-server/System.Net.Http-server.nfproj
@@ -1,0 +1,141 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="15.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup Label="Globals">
+    <NanoFrameworkProjectSystemPath>$(MSBuildToolsPath)..\..\..\nanoFramework\v1.0\</NanoFrameworkProjectSystemPath>
+  </PropertyGroup>
+  <Import Project="$(NanoFrameworkProjectSystemPath)NFProjectSystem.Default.props" Condition="Exists('$(NanoFrameworkProjectSystemPath)NFProjectSystem.Default.props')" />
+  <PropertyGroup>
+    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
+    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
+    <ProjectTypeGuids>{11A8DD76-328B-46DF-9F39-F559912D0360};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
+    <ProjectGuid>3b91a15a-40b1-4d57-93f2-eb96af0e7949</ProjectGuid>
+    <OutputType>Library</OutputType>
+    <AppDesignerFolder>Properties</AppDesignerFolder>
+    <FileAlignment>512</FileAlignment>
+    <RootNamespace>System.Net.Http</RootNamespace>
+    <AssemblyName>System.Net.Http</AssemblyName>
+    <TargetFrameworkVersion>v1.0</TargetFrameworkVersion>
+    <DocumentationFile>bin\$(Configuration)\System.Net.Http.xml</DocumentationFile>
+  </PropertyGroup>
+  <PropertyGroup>
+    <SignAssembly>true</SignAssembly>
+  </PropertyGroup>
+  <PropertyGroup>
+    <AssemblyOriginatorKeyFile>..\key.snk</AssemblyOriginatorKeyFile>
+  </PropertyGroup>
+  <PropertyGroup>
+    <DelaySign>false</DelaySign>
+  </PropertyGroup>
+  <Import Project="$(NanoFrameworkProjectSystemPath)NFProjectSystem.props" Condition="Exists('$(NanoFrameworkProjectSystemPath)NFProjectSystem.props')" />
+  <ItemGroup>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net.AuthenticationType.cs">
+      <Link>Http\System.Net.AuthenticationType.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net.Http.Constants.cs">
+      <Link>Http\System.Net.Http.Constants.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net.HttpListener.cs">
+      <Link>Http\System.Net.HttpListener.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net.HttpListenerContext.cs">
+      <Link>Http\System.Net.HttpListenerContext.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net.HttpListenerRequest.cs">
+      <Link>Http\System.Net.HttpListenerResponse.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net.HttpListenerResponse.cs">
+      <Link>
+      </Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net.HttpStatusCode.cs">
+      <Link>Http\System.Net.HttpStatusCode.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net.HttpVersion.cs">
+      <Link>Http\System.Net.HttpVersion.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net.Internal.cs">
+      <Link>Http\System.Net.Internal.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net.iwebproxy.cs">
+      <Link>Http\System.Net.iwebproxy.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net.NetworkCredential.cs">
+      <Link>Http\System.Net.NetworkCredential.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net.ProtocolViolationException.cs">
+      <Link>Http\System.Net.ProtocolViolationException.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net.WebException.cs">
+      <Link>Http\System.Net.WebException.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net.WebHeaders.cs">
+      <Link>Http\System.Net.WebHeaders.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net.WebResponse.cs">
+      <Link>\Http\System.Net.WebResponse.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net.WebStatus.cs">
+      <Link>Http\System.Net.WebStatus.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net._HeaderInfo.cs">
+      <Link>Http\System.Net._HeaderInfo.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net._HeaderInfoTable.cs">
+      <Link>Http\System.Net._HeaderInfoTable.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net._HttpDateParse.cs">
+      <Link>Http\System.Net._HttpDateParse.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net._InputNetworkStreamWrapper.cs">
+      <Link>Http\System.Net._InputNetworkStreamWrapper.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net._OutputNetworkStreamWrapper.cs">
+      <Link>Http\System.Net._OutputNetworkStreamWrapper.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Net._ValidationHelper.cs">
+      <Link>Http\System.Net._ValidationHelper.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Http\System.Uri.cs">
+      <Link>Http\System.Uri.cs</Link>
+    </Compile>
+    <Compile Include="..\nanoFramework.System.Net.Http\Properties\AssemblyInfo.cs">
+      <Link>Properties\AssemblyInfo.cs</Link>
+    </Compile>
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="..\key.snk" />
+    <None Include="\nanoFramework.System.Net.Http\packages.config" >
+      <Link>packages.config</Link>
+    </None>
+  </ItemGroup>
+  <ItemGroup>
+    <Reference Include="mscorlib, Version=1.1.1.7, Culture=neutral, PublicKeyToken=c07d481e9758c731">
+      <HintPath>..\packages\nanoFramework.CoreLibrary.1.1.1\lib\mscorlib.dll</HintPath>
+      <Private>True</Private>
+    </Reference>
+    <Reference Include="nanoFramework.Runtime.Events, Version=1.0.5.0, Culture=neutral, PublicKeyToken=c07d481e9758c731">
+      <HintPath>..\packages\nanoFramework.Runtime.Events.1.0.5-preview-008\lib\nanoFramework.Runtime.Events.dll</HintPath>
+      <Private>True</Private>
+    </Reference>
+    <Reference Include="System.Net, Version=1.0.6.0, Culture=neutral, PublicKeyToken=c07d481e9758c731">
+      <HintPath>..\packages\nanoFramework.System.Net.1.0.6-preview-011\lib\System.Net.dll</HintPath>
+      <Private>True</Private>
+    </Reference>
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Http\" />
+  </ItemGroup>
+  <Import Project="$(NanoFrameworkProjectSystemPath)NFProjectSystem.CSharp.targets" Condition="Exists('$(NanoFrameworkProjectSystemPath)NFProjectSystem.CSharp.targets')" />
+  <ProjectExtensions>
+    <ProjectCapabilities>
+      <ProjectConfigurationsDeclaredAsItems />
+    </ProjectCapabilities>
+  </ProjectExtensions>
+  <Import Project="..\packages\Nerdbank.GitVersioning.3.0.6-beta\build\Nerdbank.GitVersioning.targets" Condition="Exists('..\packages\Nerdbank.GitVersioning.3.0.6-beta\build\Nerdbank.GitVersioning.targets')" />
+  <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
+    <PropertyGroup>
+      <ErrorText>This project references NuGet package(s) that are missing on this computer. Enable NuGet Package Restore to download them.  For more information, see http://go.microsoft.com/fwlink/?LinkID=322105.The missing file is {0}.</ErrorText>
+    </PropertyGroup>
+    <Error Condition="!Exists('..\packages\Nerdbank.GitVersioning.3.0.6-beta\build\Nerdbank.GitVersioning.targets')" Text="$([System.String]::Format('$(ErrorText)', '..\packages\Nerdbank.GitVersioning.3.0.6-beta\build\Nerdbank.GitVersioning.targets'))" />
+  </Target>
+</Project>
