+++ source/nanoFramework.System.Net.Http/Http/System.Net.Internal.WebRequestPrefixElement.cs
@@ -1,0 +1,24 @@
+//
+// Copyright (c) 2018 The nanoFramework project contributors
+// Portions Copyright (c) Microsoft Corporation.  All rights reserved.
+// See LICENSE file in the project root for full license information.
+//
+
+namespace System.Net
+{
+    internal class WebRequestPrefixElement
+    {
+        public string Prefix;
+        public IWebRequestCreate Creator;
+
+        public WebRequestPrefixElement(string P, IWebRequestCreate C)
+        {
+            Prefix = P;
+            Creator = C;
+        }
+
+    } // class PrefixListElement
+
+} // namespace System.Net
+
+
